[Refactoring ⚙️] comment state 관련 로직 훅으로 분리
